+++ frontend/wailsjs/go/desktop/App.js
@@ -526,6 +526,10 @@ export function ReadFileAsDataURL(arg1) {
   return window['go']['desktop']['App']['ReadFileAsDataURL'](arg1);
 }
 
+export function ReadImageAsDataURL(arg1) {
+  return window['go']['desktop']['App']['ReadImageAsDataURL'](arg1);
+}
+
 export function Rebase(arg1) {
   return window['go']['desktop']['App']['Rebase'](arg1);
 }